2.2

+++ app/code/Magento/Config/Model/Config/Structure/Mapper/Sorting.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright © 2015 Magento. All rights reserved.
+ * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
 
@@ -9,6 +9,10 @@
  */
 namespace Magento\Config\Model\Config\Structure\Mapper;
 
+/**
+ * @api
+ * @since 100.0.2
+ */
 class Sorting extends \Magento\Config\Model\Config\Structure\AbstractMapper
 {
     /**
@@ -51,11 +55,11 @@ class Sorting extends \Magento\Config\Model\Config\Structure\AbstractMapper
     {
         $sortIndexA = 0;
         if ($this->_hasValue('sortOrder', $elementA)) {
-            $sortIndexA = intval($elementA['sortOrder']);
+            $sortIndexA = floatval($elementA['sortOrder']);
         }
         $sortIndexB = 0;
         if ($this->_hasValue('sortOrder', $elementB)) {
-            $sortIndexB = intval($elementB['sortOrder']);
+            $sortIndexB = floatval($elementB['sortOrder']);
         }
 
         if ($sortIndexA == $sortIndexB) {
